Add not found exception in doctrine infrastructure

+++ src/Infrastructure/Persistence/Cycle/DoctrineCycleRepository.php
@@ -4,8 +4,11 @@
 namespace App\Infrastructure\Persistence\Cycle;
 
 
+use App\Domain\Container\ContainerNotFoundException;
 use App\Domain\Cycle\Cycle;
+use App\Domain\Cycle\CycleNotFoundException;
 use App\Domain\Cycle\CycleRepository;
+use App\Domain\Microclimate\MicroclimateNotFoundException;
 use App\Domain\User\UserNotFoundException;
 use Cassandra\Date;
 use Psr\Container\ContainerInterface;
@@ -33,12 +36,19 @@ class DoctrineCycleRepository implements CycleRepository
 
     public function findById(int $id_cycle): Cycle
     {
-        return $this->entityManager->find("App\Domain\Cycle\Cycle",$id_cycle);
+        $cycle = $this->entityManager->find("App\Domain\Cycle\Cycle",$id_cycle);
+        if($cycle===null){
+            throw new CycleNotFoundException();
+        }
+        return $cycle;
     }
 
     public function createCycle(int $id_user, array $params): Cycle
     {
         $user = $this->entityManager->find("App\Domain\User\User",$id_user);
+        if($user===null){
+            throw new UserNotFoundException();
+        }
         $start_date = new \DateTime($params["start_date"]);
         $estimated_date = new \DateTime($params["estimated_date"]);
         $finish_date = new \DateTime($params["finish_date"]);
@@ -52,6 +62,9 @@ class DoctrineCycleRepository implements CycleRepository
     public function deleteCycle(int $id): bool
     {
         $cycle = $this->entityManager->find("App\Domain\Cycle\Cycle",$id);
+        if($cycle===null){
+            throw new CycleNotFoundException();
+        }
         $this->entityManager->remove($cycle);
         $this->entityManager->flush();
         return true;
@@ -60,6 +73,9 @@ class DoctrineCycleRepository implements CycleRepository
     public function updateCycle(int $id, array $params): Cycle
     {
         $cycle = $this->entityManager->find("App\Domain\Cycle\Cycle",$id);
+        if($cycle===null){
+            throw new CycleNotFoundException();
+        }
         if(isset($params["estimated_date"])){
             $estimated_date = new \DateTime($params["estimated_date"]);
             $cycle->setEstimatedDate($estimated_date);
@@ -88,8 +104,17 @@ class DoctrineCycleRepository implements CycleRepository
     public function createCycleContainerMicroclimate(int $id_user, int $id_container, int $id_microclimate, array $params): Cycle
     {
         $user = $this->entityManager->find("App\Domain\User\User",$id_user);
+        if($user===null){
+            throw new UserNotFoundException();
+        }
         $container = $this->entityManager->find("App\Domain\Container\Container",$id_container);
+        if($container===null){
+            throw new ContainerNotFoundException();
+        }
         $microclimate = $this->entityManager->find("App\Domain\Microclimate\Microclimate",$id_microclimate);
+        if($microclimate===null){
+            throw new MicroclimateNotFoundException();
+        }
         $start_date = new \DateTime($params["start_date"]);
         $estimated_date = new \DateTime($params["estimated_date"]);
         $finish_date = new \DateTime($params["finish_date"]);
